final implementation middleware

+++ app/Http/Controllers/Backsite/DoctorController.php
@@ -13,7 +13,8 @@ use Symfony\Component\HttpFoundation\Response;
 use App\Http\Requests\Doctor\StoreDoctorRequest;
 use App\Http\Requests\Doctor\UpdateDoctorRequest;
 
-//autentikasi
+// use everything here
+use Gate;
 use Auth;
 
 use App\Models\Operational\Doctor;
@@ -80,6 +81,9 @@ class DoctorController extends Controller
      */
     public function show(Doctor $doctor)
     {
+        
+        abort_if(Gate::denies('doctor_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         return view('pages.backsite.operational.doctor.show', compact('doctor'));
     }
 
@@ -91,6 +95,10 @@ class DoctorController extends Controller
      */
     public function edit(Doctor $doctor)
     {
+
+        abort_if(Gate::denies('doctor_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
+        // for select2 = ascending a to z
         $specialist = Specialist::orderBy('name', 'asc')->get();
 
         return view('pages.backsite.operational.doctor.edit', compact('doctor', 'specialist'));
@@ -125,9 +133,12 @@ class DoctorController extends Controller
     public function destroy(Doctor $doctor)
     {
  
-       $doctor->forceDelete(); 
+        abort_if(Gate::denies('doctor_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
-       alert()->success('Success Message', 'Successfully deleted data');
-       return back();
+        $doctor->forceDelete(); 
+
+        alert()->success('Success Message', 'Successfully deleted data');
+        
+        return back();
     }
 }
